entity/orbit: fixed ring generation
The ring generation had a minor oversight wherein the rings started generating from the centre of it's owning planet; when they are supposed to start generating from the surface. I've fixed that.
This also reveals more of the rings as none are now hidden under the surface; so I've reduced the ring count.