entity/player: introduced player state API
The player entity now provides a public method for inspecting it's state: `Player::is(PlayerState const&)`; it takes in a player state and evaluates whether the player is in that state. In practice, this is very easy to use; for example: `Player::is(PlayerState::InsideSmoothingRing)` will evaluate whether the player is inside the smoothing ring of the active planet's orbit.

It's implementation is very bad and poc right now; but it works and I'll refine it later.