entity/player: clamp velocity when outside orbit
Prevents the player from racking up infinite acceleration by timing orbit toggles.

Since the orbit "accelerates" the player into a circular path; if you turn it off right after getting accelerated; the player gets accelerated in the forwards direction instead. Repeat this a few times and you'd be zipping around the map.

This change introduces a clamp on the velocity; so the player can still use this mechanic; but only upto a reasonable extent.