entity/player: only clamp velocity when drifting
Earlier, we were clamping the velocity of the player when it was anywhere outside the orbit of a planet; this caused a bug where planets would not be able to "pull in" the player because the inward acceleration applied on the player would get clamped. This looked very strange and overall made the game less fun.

Now we only clamp when `Player::is(FarOutsideOrbit)` that ensures we do not interfere with the planet's gravity.