graphics/particles: impl basic particle emitter with demo
Simple particle emitter system which blocks the main game ::update() loop on particle emission start; manages the particles and automatically unblocks the update loop when all particles expire. Right now a demo "explosion animation" is played on player collision; it is not final but serves to show how this particle emitter works.