core/nav: converted to singleton
I did not want to do this at first since we already have quite a few singletons; but navigation is way too broad and way too many components need this information to not do this. So I finally caved.

Minor issue: Since there is no way to have a nullable reference in C++, atleast afaik, I am using a raw pointer to player which is provided through `Navigation::bind`, as the player instance is not available when the navigation singleton will be instantiated. Not a fan but it's fine for now.