core/nav: redesigned API
Navigation now manages and exposes a `NavigationContext`; which is calculated exactly ONCE on every frame and then can be polled by any component in the project.

The Navigation::make_context() method is a little ugly; but I couldn't find a better way to find all the information we need for the navigation context using a single loop. Could refactor later.

None of the users of Navigation have been updated; so the code won't compile right now.