core/nav/ctx: improved previous planet selection logic
The earlier impl checked for the strict `InTargetOrbit` state; which in many scenarios is not met even when the player is in a reasonable orbit a little bit further away from the target. This often caused the player to respawn a  couple planets back, esp if they switched planets quickly. The new more relaxed `InStableOrbit` state seems to work significantly better.